backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
openai==1.12.0